+++ client/.gitignore
@@ -4,7 +4,6 @@
 /node_modules
 /.pnp
 .pnp.js
-/.idea
 
 # testing
 /coverage
