--- growingup/src/components/CategoryList/CategoryList.css
@@ -1,10 +1,0 @@
-
-
-
-.mainCategoryList {
-    margin: -3px;
-    /* position: absolute; */
-    position: fixed;
-    margin-top: auto;
-    padding-top: 65px;
-}
