--- growingup/src/components/Forums/ForumForm.css
@@ -1,5 +1,0 @@
-.forumForm {
-    margin: 100px;
-    border: solid 10px black;
-    margin-top: 82px;
-}
