--- growingup/src/components/Forums/ForumPost.css
@@ -1,9 +1,0 @@
-.forumCard {
-    padding: 25px;
-    margin: 368px;
-    margin-top: 1em;
-}
-
-Col {
-    display:flex;
-}
