--- growingup/src/components/Forums/ForumPost.js
@@ -1,44 +1,0 @@
-import React from 'react';
-import './ForumPost.css';
-// import { Card, CardImg, CardText, CardBody, CardLink,
-//   CardTitle, CardSubtitle } from 'reactstrap';
-
-// const Example = (props) => {
-//   return (
-//     <div>
-//       <Card className="forumCard">
-//         <CardBody>
-//           <CardTitle>Card title</CardTitle>
-//           <CardSubtitle>Card subtitle</CardSubtitle>
-//         </CardBody>
-//         <img width="100%" src="https://placeholdit.imgix.net/~text?txtsize=33&txt=318%C3%97180&w=318&h=180" alt="Card image cap" />
-//         <CardBody>
-//           <CardText>Some quick example text to build on the card title and make up the bulk of the card's content.</CardText>
-//           <CardLink href="#">Card Link</CardLink>
-//           <CardLink href="#">Another Link</CardLink>
-//         </CardBody>
-//       </Card>
-//     </div>
-//   );
-// };
-
-// export default Example;
-
-// import React from 'react';
-import { Card, Button, CardTitle, CardText, Row, Col } from 'reactstrap';
-
-const Example = (props) => {
-  return (
-    <Row>
-      <Col sm="6">
-        <Card body>
-          <CardTitle>Special Title Treatment</CardTitle>
-          <CardText>With supporting text below as a natural lead-in to additional content.</CardText>
-          <Button>Go somewhere</Button>
-        </Card>
-      </Col>
-    </Row>
-  );
-};
-
-export default Example;
