--- growingup/src/components/LogIn/LogIn.css
@@ -1,26 +1,0 @@
-.logInForm {
-    color: whitesmoke;
-    background-color: red;
-    padding: 50px;
-    margin: 100px;
-}
-
-#formGroup1, #formGroup2 {
- line-height: 30px;
- padding-bottom: 20px;
-}
-
-.logInSubmitButton {
-    padding-top: 1px;
-    padding-bottom: 2px;
-    margin-bottom:10px;
-    background: blue;
-}
-
-h2 {
-    padding-bottom: 0.5em;
-}
-
-.signUpLinkButton {
-    padding-top: 10%;
-}
