--- growingup/src/index.css
@@ -1,30 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-
-#masthead {
-  z-index: 99;
-  width: 100%;
-  /* margin: 0 0 30px; */
-  overflow: auto;
-  opacity: 0.6;
-  position: fixed;
-  background-color: #0f292f;
-}
-/* nav a {
-  color:white;
-} */
-
-  
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
-}
